final_pkg/company_nemo_guardrails/requirements.txt
pytest>=8.0.0
pyyaml>=6.0.1
openai>=1.0.0
nemoguardrails>=0.21.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
